Gjør Turplan til en app som får folk ut: turforslag, filtre og dagbok
Appen åpner nå på turforslag i stedet for et tomt kart, og alt det tekniske
er flyttet ut av veien.

Finn tur
- Henter ekte, navngitte turer fra Turrutebasen og setter løse rutesegmenter
  sammen til hele turer (Preikestolen, Tromsdalstinden, Fløyveien …).
- Filtrer på lengde, vanskegrad, rundtur, merking og rutetype. Filtrene
  virker umiddelbart, uten nye nettverkskall.
- Hvert kort viser lengde, stigning, tidsbruk og et lite høydeprofil-riss.
  Ett trykk laster hele turen inn i planleggeren.
- «Overrask meg» plukker en tilfeldig tur når man ikke klarer å bestemme seg.

Turdagbok
- Marker en tur som gått, så samler den seg opp med kilometer og høydemeter.
- Fjorten merker å samle, og sammenligninger som forteller hvor langt det er:
  «du har klatret 1,4 × Galdhøpiggen».
- Alt ligger lokalt i nettleseren.

Ryddigere grensesnitt
- Tre faner: Finn tur, Turen, Dagbok. Vær og sikkerhet ligger som seksjoner
  i turen, ikke som egne faner.
- Marsjfart og starttidspunkt er synlige; underlag, sekkevekt, pauser, følg
  sti og GPX ligger under «Avansert».
- Kartlagene er flyttet til en knapp på selve kartet.
- Bekreftelsesdialogen ved tømming er byttet mot en angreknapp i meldingen.

Rettelser funnet av testene
- Kartet vokste til 11 000 piksler når turlista ble lang: rutenettraden
  manglet minmax(0, 1fr).
- [hidden] ble overstyrt av klasser som setter display, så tegneknappene
  vistes uten rute.
- Escape lukket ikke kartlagsvinduet når fokus sto inni det.

+++ src/js/util.js
@@ -69,11 +69,9 @@ export function formatClock(date) {
   return new Intl.DateTimeFormat('nb-NO', { hour: '2-digit', minute: '2-digit' }).format(date);
 }
 
-/** onsdag 13. sep. */
+/** onsdag 13. september */
 export function formatDay(date) {
-  return new Intl.DateTimeFormat('nb-NO', { weekday: 'long', day: 'numeric', month: 'short' }).format(
-    date,
-  );
+  return new Intl.DateTimeFormat('nb-NO', { weekday: 'long', day: 'numeric', month: 'long' }).format(date);
 }
 
 /** 59.8542° N, 8.6492° Ø – desimalgrader er det Kartverket bruker. */
